+++ incapsula/resource_site_test.go
@@ -106,7 +106,6 @@ func testAccCheckIncapsulaSiteConfigBasic(domain string) string {
 	return fmt.Sprintf(`
 		resource "incapsula_site" "testacc-terraform-site" {
 			domain = "%s"
-			data_storage_region = "US"
 		}`,
 		domain,
 	)